refactor: Move utility functions to common module
