+++ src/asm.rs
@@ -1,4 +1,7 @@
-use crate::common::{Function, Identifier, IdentifierTable, Intrinsic, Op, OpType, Segment};
+use crate::common::{
+    Function, Identifier, IdentifierTable, Intrinsic, Op, OpType, Segment, get_related_done_id,
+    get_related_fi_id, get_related_while_id,
+};
 
 pub fn generate_assembly_code(segments: &[Segment], identifier_table: &IdentifierTable) -> String {
     let mut asm_blocks = Vec::new();
@@ -224,57 +227,6 @@ fn get_asm_intrinsic(intrinsic: &Intrinsic) -> String {
         Intrinsic::Syscall6 => get_asm_syscall(6),
         _ => todo!(),
     }
-}
-
-fn get_related_fi_id(op: &Op, function: &Function) -> Option<usize> {
-    assert!(op.ty == OpType::Then);
-
-    let op_index = function.ops.binary_search_by_key(&op.id, |op| op.id).ok()?;
-    let mut nested_ifs = 0;
-
-    for other_op in &function.ops[op_index + 1..] {
-        match other_op.ty {
-            OpType::Fi if nested_ifs == 0 => return Some(other_op.id),
-            OpType::Fi => nested_ifs -= 1,
-            OpType::If => nested_ifs += 1,
-            _ => {}
-        }
-    }
-    None
-}
-
-fn get_related_done_id(op: &Op, function: &Function) -> Option<usize> {
-    assert!(op.ty == OpType::Break || op.ty == OpType::Do);
-
-    let op_index = function.ops.binary_search_by_key(&op.id, |op| op.id).ok()?;
-    let mut nested_whiles = 0;
-
-    for other_op in &function.ops[op_index + 1..] {
-        match other_op.ty {
-            OpType::Done if nested_whiles == 0 => return Some(other_op.id),
-            OpType::Done => nested_whiles -= 1,
-            OpType::While => nested_whiles += 1,
-            _ => {}
-        }
-    }
-    None
-}
-
-fn get_related_while_id(op: &Op, function: &Function) -> Option<usize> {
-    assert!(op.ty == OpType::Continue || op.ty == OpType::Done);
-
-    let op_index = function.ops.binary_search_by_key(&op.id, |op| op.id).ok()?;
-    let mut nested_whiles = 0;
-
-    for other_op in function.ops[..op_index].iter().rev() {
-        match other_op.ty {
-            OpType::While if nested_whiles == 0 => return Some(other_op.id),
-            OpType::Done => nested_whiles += 1,
-            OpType::While => nested_whiles -= 1,
-            _ => {}
-        }
-    }
-    None
 }
 
 fn get_asm_push_bind(op: &Op, function: &Function) -> String {
